% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binToDec.R
\name{binToDec}
\alias{binToDec}
\title{Convert binary to decimal}
\usage{
binToDec(x)
}
\arguments{
\item{x}{binary sequence}
}
\value{
value decoded into decimal system
}
\description{
Create an auxiliary function for decoding a number in binary system into a number in decimal system
}
