% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{comparePopulPlot}
\alias{comparePopulPlot}
\title{Compare first and last iter}
\usage{
comparePopulPlot()
}
\value{
ggplot
}
\description{
Generates comparison of populations in the first and the last generation
}
