% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load data}
\usage{
loadData(x, start, end, freq)
}
\arguments{
\item{x}{tick of required stock (ex. "PKO")}

\item{start}{start date (format = yyyymmdd)}

\item{end}{end date (format = yyyymmdd)}

\item{freq}{monthly('m')/daily('d') data}
}
\value{
data.frame with share prices
}
\description{
Function loads the required stock monthly data for tick *x* from https://stooq.pl.
}
