% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioDetails.R
\name{portfolioReturn}
\alias{portfolioReturn}
\title{Show portfolio return}
\usage{
portfolioReturn(x, num_genes, expectedReturn. = expectedReturn,
  num_stocks. = num_stocks)
}
\arguments{
\item{x}{single portfolio (sequence of bits)}

\item{num_genes}{number of bits encoding single stock}

\item{num_stocks.}{number of stocks in analysis}
}
\value{
Value for expected return
}
\description{
Function decodes sequence of bits in *x* and returns expected return for a given portfolio.
}
