% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioDetails.R
\name{portfolioRisk}
\alias{portfolioRisk}
\title{Show portfolio risk}
\usage{
portfolioRisk(x, num_genes, num_stocks. = num_stocks)
}
\arguments{
\item{x}{single portfolio (sequence of bits)}

\item{num_genes}{number of bits encoding single stock}

\item{num_stocks.}{number of stocks in analysis}
}
\value{
Value of risk
}
\description{
Function decodes sequence of bits in *x* and returns value of risk for a given portfolio.
}
