% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Prepare data}
\usage{
prepareData(data)
}
\arguments{
\item{data}{raw dataset loaded with @loadData function}
}
\value{
data - modified dataset

chng - matrix with percentage changes in stock prices

cor, cov, expectedReturn, risk - measures of stocks under analysis

num_stocks - number of stocks in analysis
}
\description{
Function prepares all the necessary data for further analysis. Requires the dataset loaded with @loadData function. Function assigns the Values
into .GlobalEnv.
}
