% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioDetails.R
\name{showPortfolio}
\alias{showPortfolio}
\title{Show portfolio structure}
\usage{
showPortfolio(x, num_genes, num_stocks. = num_stocks)
}
\arguments{
\item{x}{single portfolio (sequence of bits)}

\item{num_genes}{number of bits encoding single stock}

\item{num_stocks.}{number of stocks in analysis}
}
\value{
Vector with portfolio structure.
}
\description{
Function decodes sequence of bits in *x* and returns the portfolio structure.
}
