% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleAdapt.R
\name{singleAdapt}
\alias{singleAdapt}
\title{Count speciman adaptation}
\usage{
singleAdapt(x, num_genes, expectedReturn. = expectedReturn,
  num_stocks. = num_stocks)
}
\arguments{
\item{x}{speciman, which is a sequence of bits}

\item{num_genes}{number of bits encoding single stock}

\item{expectedReturn.}{a vector with expected return rates for analysed stocks}

\item{num_stocks.}{number of stocks in analysis}
}
\value{
Value of adaptation
}
\description{
Create an auxiliary function for counting a single speciman adaptation (single portfolio quality)
Function takes *x* which is a single specimen (sequence of bits) and returns a value of its adaptation. The adaptation value is the
expected portfolio return / portfolio risk.
}
